Modularize postgresql adapter
