+++ activerecord/lib/active_record/connection_adapters/postgresql/referential_integrity.rb
@@ -1,0 +1,22 @@
+module ActiveRecord
+  module ConnectionAdapters
+    class PostgreSQLAdapter < AbstractAdapter
+      module ReferentialIntegrity
+        def supports_disable_referential_integrity? #:nodoc:
+          true
+        end
+
+        def disable_referential_integrity #:nodoc:
+          if supports_disable_referential_integrity? then
+            execute(tables.collect { |name| "ALTER TABLE #{quote_table_name(name)} DISABLE TRIGGER ALL" }.join(";"))
+          end
+          yield
+        ensure
+          if supports_disable_referential_integrity? then
+            execute(tables.collect { |name| "ALTER TABLE #{quote_table_name(name)} ENABLE TRIGGER ALL" }.join(";"))
+          end
+        end
+      end
+    end
+  end
+end
